Download 2020 Block Assignment Files for a State

From the Census: "The Block Assignment Files (BAFs) are among the geographic products that the Census Bureau provides to states and other data users containing the small area census data necessary for legislative redistricting. The BAFs contain Census tabulation block codes and geographic area codes for a specific geographic entity type."

pl_get_baf(abbr, geographies = NULL, cache_to = NULL, refresh = FALSE)

Arguments

abbr

the state abbreviation to get the BAF for

geographies

the geographies to get. Defaults to all available.

cache_to

the file name, if any, to cache the results to (as an RDS). If a file exists and refresh=FALSE, will read BAF from this file.

refresh

if TRUE, force a re-download of the BAF data.

Value

A list of tibbles, one for each available BAF geography.

Examples

pl_get_baf("RI")
#> $INCPLACE_CDP
#> # A tibble: 25,649 × 2
#>    BLOCKID         PLACEFP
#>    <chr>           <chr>  
#>  1 440010301001000 NA     
#>  2 440010301001001 NA     
#>  3 440010301001002 NA     
#>  4 440010301001003 NA     
#>  5 440010301001004 NA     
#>  6 440010301001005 NA     
#>  7 440010301001006 NA     
#>  8 440010301001007 NA     
#>  9 440010301001008 NA     
#> 10 440010301001009 NA     
#> # … with 25,639 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $MCD
#> # A tibble: 25,649 × 3
#>    BLOCKID         COUNTYFP COUSUBFP
#>    <chr>           <chr>    <chr>   
#>  1 440010301001000 001      05140   
#>  2 440010301001001 001      05140   
#>  3 440010301001002 001      05140   
#>  4 440010301001003 001      05140   
#>  5 440010301001004 001      05140   
#>  6 440010301001005 001      05140   
#>  7 440010301001006 001      05140   
#>  8 440010301001007 001      05140   
#>  9 440010301001008 001      05140   
#> 10 440010301001009 001      05140   
#> # … with 25,639 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $VTD
#> # A tibble: 25,649 × 3
#>    BLOCKID         COUNTYFP DISTRICT
#>    <chr>           <chr>    <chr>   
#>  1 440010301001000 001      440101  
#>  2 440010301001001 001      440101  
#>  3 440010301001002 001      440101  
#>  4 440010301001003 001      440101  
#>  5 440010301001004 001      440101  
#>  6 440010301001005 001      440101  
#>  7 440010301001006 001      440101  
#>  8 440010301001007 001      440101  
#>  9 440010301001008 001      440101  
#> 10 440010301001009 001      440101  
#> # … with 25,639 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $CD
#> # A tibble: 25,649 × 2
#>    BLOCKID         DISTRICT
#>    <chr>           <chr>   
#>  1 440010301001000 01      
#>  2 440010301001001 01      
#>  3 440010301001002 01      
#>  4 440010301001003 01      
#>  5 440010301001004 01      
#>  6 440010301001005 01      
#>  7 440010301001006 01      
#>  8 440010301001007 01      
#>  9 440010301001008 01      
#> 10 440010301001009 01      
#> # … with 25,639 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $SLDU
#> # A tibble: 25,649 × 2
#>    BLOCKID         DISTRICT
#>    <chr>           <chr>   
#>  1 440010301001000 032     
#>  2 440010301001001 032     
#>  3 440010301001002 032     
#>  4 440010301001003 032     
#>  5 440010301001004 032     
#>  6 440010301001005 032     
#>  7 440010301001006 032     
#>  8 440010301001007 032     
#>  9 440010301001008 032     
#> 10 440010301001009 032     
#> # … with 25,639 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $SLDL
#> # A tibble: 25,649 × 2
#>    BLOCKID         DISTRICT
#>    <chr>           <chr>   
#>  1 440010301001000 066     
#>  2 440010301001001 066     
#>  3 440010301001002 066     
#>  4 440010301001003 066     
#>  5 440010301001004 066     
#>  6 440010301001005 066     
#>  7 440010301001006 066     
#>  8 440010301001007 066     
#>  9 440010301001008 066     
#> 10 440010301001009 066     
#> # … with 25,639 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $SDSEC
#> # A tibble: 25,649 × 2
#>    BLOCKID         DISTRICT
#>    <chr>           <chr>   
#>  1 440010301001000 NA      
#>  2 440010301001001 NA      
#>  3 440010301001002 NA      
#>  4 440010301001003 NA      
#>  5 440010301001004 NA      
#>  6 440010301001005 NA      
#>  7 440010301001006 NA      
#>  8 440010301001007 NA      
#>  9 440010301001008 NA      
#> 10 440010301001009 NA      
#> # … with 25,639 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $SDELM
#> # A tibble: 25,649 × 2
#>    BLOCKID         DISTRICT
#>    <chr>           <chr>   
#>  1 440010301001000 NA      
#>  2 440010301001001 NA      
#>  3 440010301001002 NA      
#>  4 440010301001003 NA      
#>  5 440010301001004 NA      
#>  6 440010301001005 NA      
#>  7 440010301001006 NA      
#>  8 440010301001007 NA      
#>  9 440010301001008 NA      
#> 10 440010301001009 NA      
#> # … with 25,639 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $SDUNI
#> # A tibble: 25,649 × 2
#>    BLOCKID         DISTRICT
#>    <chr>           <chr>   
#>  1 440010301001000 00030   
#>  2 440010301001001 00030   
#>  3 440010301001002 00030   
#>  4 440010301001003 00030   
#>  5 440010301001004 00030   
#>  6 440010301001005 00030   
#>  7 440010301001006 00030   
#>  8 440010301001007 00030   
#>  9 440010301001008 00030   
#> 10 440010301001009 00030   
#> # … with 25,639 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $AIANNH
#> # A tibble: 25,649 × 3
#>    BLOCKID         AIANNHCE COMPTYP
#>    <chr>           <chr>    <chr>  
#>  1 440010301001000 NA       NA     
#>  2 440010301001001 NA       NA     
#>  3 440010301001002 NA       NA     
#>  4 440010301001003 NA       NA     
#>  5 440010301001004 NA       NA     
#>  6 440010301001005 NA       NA     
#>  7 440010301001006 NA       NA     
#>  8 440010301001007 NA       NA     
#>  9 440010301001008 NA       NA     
#> 10 440010301001009 NA       NA     
#> # … with 25,639 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
pl_get_baf("RI", "VTD")
#> $VTD
#> # A tibble: 25,649 × 3
#>    BLOCKID         COUNTYFP DISTRICT
#>    <chr>           <chr>    <chr>   
#>  1 440010301001000 001      440101  
#>  2 440010301001001 001      440101  
#>  3 440010301001002 001      440101  
#>  4 440010301001003 001      440101  
#>  5 440010301001004 001      440101  
#>  6 440010301001005 001      440101  
#>  7 440010301001006 001      440101  
#>  8 440010301001007 001      440101  
#>  9 440010301001008 001      440101  
#> 10 440010301001009 001      440101  
#> # … with 25,639 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#>