Download TIGER Prototype shapefiles

These prototype shapefiles correspond to the Rhode Island end-to-end Census test and the accompanying prototype P.L. 94-171 data. This function is unlikely to be useful for working with any actual decennial Census data. The corresponding tinytiger or tigris functions should be used instead.

pl_get_prototype(
  geog,
  year = 2020,
  full_state = TRUE,
  cache_to = NULL,
  clean_names = TRUE,
  refresh = FALSE
)

Arguments

geog

Geography to download data for. See details for full list.

year

year, either 2010 or 2020

full_state

whether to return the full state (TRUE) or the single county subset (FALSE)

cache_to

the file name, if any, to cache the results to (as an RDS). If a file exists and refresh=FALSE, will read from this file.

clean_names

whether to clean and rename columns

refresh

if TRUE, force a re-download of the data.

Value

An sf object containing the blocks.

Details

Current acceptable arguments to geog include:

block: block

block_group: block group

tract: tract

county: county

state: state

sld_low: state legislative district lower house

sld_up: state legislative district upper house

congressional_district: federal congressional district

place: Census place

voting_district: voting tabulation district

Examples

# \donttest{
shp <- pl_get_prototype("block")
# }