Approximately re-tally Census data under new block boundaries

Applies a block crosswalk to a table of block data using areal interpolation. That is, the fraction of land area in the overlapping region between old and new blocks is used to divide the population of the old blocks into the new.

pl_retally(d_from, crosswalk)

Arguments

d_from
The data frame to process. All numeric columns will be re-tallied. Integer columns will be re-tallied with rounding. Character columns will be preserved if constant across new block geometries.

crosswalk
The crosswalk data frame, from pl_crosswalk()

Value

A new data frame, like d_from, except with the geometry column dropped, if one exists. New geometry should be loaded, perhaps with tigris::blocks().

Details

All numeric columns will be re-tallied. Integer columns will be re-tallied with rounding. Character columns will be preserved if constant across new block geometries.

Blocks from other states will be ignored.

Examples

# \donttest{
crosswalk = pl_crosswalk("RI", 2010, 2020)
RI_2010 = pl_tidy_shp("RI", pl_url("RI", 2010), 2010)

#> Downloading: 20 kB     Downloading: 20 kB     Downloading: 28 kB     Downloading: 28 kB     Downloading: 40 kB     Downloading: 40 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 49 kB     Downloading: 84 kB     Downloading: 84 kB     Downloading: 84 kB     Downloading: 84 kB     Downloading: 120 kB     Downloading: 120 kB     Downloading: 140 kB     Downloading: 140 kB     Downloading: 150 kB     Downloading: 150 kB     Downloading: 150 kB     Downloading: 150 kB     Downloading: 170 kB     Downloading: 170 kB     Downloading: 190 kB     Downloading: 190 kB     Downloading: 190 kB     Downloading: 190 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 210 kB     Downloading: 240 kB     Downloading: 240 kB     Downloading: 240 kB     Downloading: 240 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 310 kB     Downloading: 370 kB     Downloading: 370 kB     Downloading: 370 kB     Downloading: 370 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 440 kB     Downloading: 450 kB     Downloading: 450 kB     Downloading: 450 kB     Downloading: 450 kB     Downloading: 490 kB     Downloading: 490 kB     Downloading: 500 kB     Downloading: 500 kB     Downloading: 520 kB     Downloading: 520 kB     Downloading: 570 kB     Downloading: 570 kB     Downloading: 630 kB     Downloading: 630 kB     Downloading: 640 kB     Downloading: 640 kB     Downloading: 690 kB     Downloading: 690 kB     Downloading: 690 kB     Downloading: 690 kB     Downloading: 760 kB     Downloading: 760 kB     Downloading: 760 kB     Downloading: 760 kB     Downloading: 830 kB     Downloading: 830 kB     Downloading: 890 kB     Downloading: 890 kB     Downloading: 890 kB     Downloading: 890 kB     Downloading: 960 kB     Downloading: 960 kB     Downloading: 960 kB     Downloading: 960 kB     Downloading: 1 MB     Downloading: 1 MB     Downloading: 1 MB     Downloading: 1 MB     Downloading: 1 MB     Downloading: 1 MB     Downloading: 1.1 MB     Downloading: 1.1 MB     Downloading: 1.1 MB     Downloading: 1.1 MB     Downloading: 1.2 MB     Downloading: 1.2 MB     Downloading: 1.2 MB     Downloading: 1.2 MB     Downloading: 1.2 MB     Downloading: 1.2 MB     Downloading: 1.2 MB     Downloading: 1.2 MB     Downloading: 1.3 MB     Downloading: 1.3 MB     Downloading: 1.3 MB     Downloading: 1.3 MB     Downloading: 1.3 MB     Downloading: 1.3 MB     Downloading: 1.4 MB     Downloading: 1.4 MB     Downloading: 1.5 MB     Downloading: 1.5 MB     Downloading: 1.5 MB     Downloading: 1.5 MB     Downloading: 1.5 MB     Downloading: 1.5 MB     Downloading: 1.5 MB     Downloading: 1.5 MB     Downloading: 1.5 MB     Downloading: 1.5 MB     Downloading: 1.6 MB     Downloading: 1.6 MB     Downloading: 1.8 MB     Downloading: 1.8 MB     Downloading: 1.8 MB     Downloading: 1.8 MB     Downloading: 1.9 MB     Downloading: 1.9 MB     Downloading: 1.9 MB     Downloading: 1.9 MB     Downloading: 1.9 MB     Downloading: 1.9 MB     Downloading: 2 MB     Downloading: 2 MB     Downloading: 2 MB     Downloading: 2 MB     Downloading: 2.2 MB     Downloading: 2.2 MB     Downloading: 2.2 MB     Downloading: 2.2 MB     Downloading: 2.3 MB     Downloading: 2.3 MB     Downloading: 2.3 MB     Downloading: 2.3 MB     Downloading: 2.4 MB     Downloading: 2.4 MB     Downloading: 2.4 MB     Downloading: 2.4 MB     Downloading: 2.5 MB     Downloading: 2.5 MB     Downloading: 2.5 MB     Downloading: 2.5 MB     Downloading: 2.5 MB     Downloading: 2.5 MB     Downloading: 2.5 MB     Downloading: 2.5 MB     Downloading: 2.5 MB     Downloading: 2.5 MB     Downloading: 2.6 MB     Downloading: 2.6 MB     Downloading: 2.6 MB     Downloading: 2.6 MB     Downloading: 2.6 MB     Downloading: 2.6 MB     Downloading: 2.7 MB     Downloading: 2.7 MB     Downloading: 2.7 MB     Downloading: 2.7 MB     Downloading: 2.7 MB     Downloading: 2.7 MB     Downloading: 2.8 MB     Downloading: 2.8 MB     Downloading: 2.8 MB     Downloading: 2.8 MB     Downloading: 2.9 MB     Downloading: 2.9 MB     Downloading: 2.9 MB     Downloading: 2.9 MB     Downloading: 2.9 MB     Downloading: 2.9 MB     Downloading: 3 MB     Downloading: 3 MB     Downloading: 3.1 MB     Downloading: 3.1 MB     Downloading: 3.1 MB     Downloading: 3.1 MB     Downloading: 3.1 MB     Downloading: 3.1 MB     Downloading: 3.1 MB     Downloading: 3.1 MB     Downloading: 3.2 MB     Downloading: 3.2 MB     Downloading: 3.2 MB     Downloading: 3.2 MB     Downloading: 3.2 MB     Downloading: 3.2 MB     Downloading: 3.3 MB     Downloading: 3.3 MB     Downloading: 3.3 MB     Downloading: 3.3 MB     Downloading: 3.4 MB     Downloading: 3.4 MB     Downloading: 3.4 MB     Downloading: 3.4 MB     Downloading: 3.4 MB     Downloading: 3.4 MB     Downloading: 3.5 MB     Downloading: 3.5 MB     Downloading: 3.6 MB     Downloading: 3.6 MB     Downloading: 3.6 MB     Downloading: 3.6 MB     Downloading: 3.6 MB     Downloading: 3.6 MB     Downloading: 3.6 MB     Downloading: 3.6 MB     Downloading: 3.6 MB     Downloading: 3.6 MB     Downloading: 3.6 MB     Downloading: 3.6 MB     Downloading: 3.7 MB     Downloading: 3.7 MB     Downloading: 3.8 MB     Downloading: 3.8 MB     Downloading: 3.8 MB     Downloading: 3.8 MB     Downloading: 3.8 MB     Downloading: 3.8 MB     Downloading: 3.9 MB     Downloading: 3.9 MB     Downloading: 4 MB     Downloading: 4 MB     Downloading: 4 MB     Downloading: 4 MB     Downloading: 4 MB     Downloading: 4 MB     Downloading: 4 MB     Downloading: 4 MB     Downloading: 4.1 MB     Downloading: 4.1 MB     Downloading: 4.1 MB     Downloading: 4.1 MB     Downloading: 4.1 MB     Downloading: 4.1 MB     Downloading: 4.1 MB     Downloading: 4.1 MB     Downloading: 4.2 MB     Downloading: 4.2 MB     Downloading: 4.2 MB     Downloading: 4.2 MB     Downloading: 4.2 MB     Downloading: 4.2 MB     Downloading: 4.3 MB     Downloading: 4.3 MB     Downloading: 4.3 MB     Downloading: 4.3 MB     Downloading: 4.3 MB     Downloading: 4.3 MB     Downloading: 4.4 MB     Downloading: 4.4 MB     Downloading: 4.4 MB     Downloading: 4.4 MB     Downloading: 4.4 MB     Downloading: 4.4 MB     Downloading: 4.5 MB     Downloading: 4.5 MB     Downloading: 4.5 MB     Downloading: 4.5 MB     Downloading: 4.5 MB     Downloading: 4.5 MB     Downloading: 4.5 MB     Downloading: 4.5 MB     Downloading: 4.5 MB     Downloading: 4.5 MB     Downloading: 4.5 MB     Downloading: 4.5 MB     Downloading: 4.5 MB     Downloading: 4.5 MB     Downloading: 4.5 MB     Downloading: 4.5 MB     Downloading: 4.6 MB     Downloading: 4.6 MB     Downloading: 4.6 MB     Downloading: 4.6 MB     Downloading: 4.7 MB     Downloading: 4.7 MB     Downloading: 4.7 MB     Downloading: 4.7 MB     Downloading: 4.7 MB     Downloading: 4.7 MB     Downloading: 4.8 MB     Downloading: 4.8 MB     Downloading: 4.9 MB     Downloading: 4.9 MB     Downloading: 4.9 MB     Downloading: 4.9 MB     Downloading: 5 MB     Downloading: 5 MB     Downloading: 5.1 MB     Downloading: 5.1 MB     Downloading: 5.1 MB     Downloading: 5.1 MB     Downloading: 5.2 MB     Downloading: 5.2 MB     Downloading: 5.2 MB     Downloading: 5.2 MB     Downloading: 5.2 MB     Downloading: 5.2 MB     Downloading: 5.2 MB     Downloading: 5.2 MB     Downloading: 5.3 MB     Downloading: 5.3 MB     Downloading: 5.3 MB     Downloading: 5.3 MB     Downloading: 5.3 MB     Downloading: 5.3 MB     Downloading: 5.3 MB     Downloading: 5.3 MB     Downloading: 5.4 MB     Downloading: 5.4 MB     Downloading: 5.4 MB     Downloading: 5.4 MB     Downloading: 5.4 MB     Downloading: 5.4 MB     Downloading: 5.4 MB     Downloading: 5.4 MB     Downloading: 5.4 MB     Downloading: 5.4 MB     Downloading: 5.4 MB     Downloading: 5.4 MB     Downloading: 5.4 MB     Downloading: 5.4 MB     Downloading: 5.5 MB     Downloading: 5.5 MB     Downloading: 5.5 MB     Downloading: 5.5 MB     Downloading: 5.6 MB     Downloading: 5.6 MB     Downloading: 5.6 MB     Downloading: 5.6 MB     Downloading: 5.6 MB     Downloading: 5.6 MB     Downloading: 5.6 MB     Downloading: 5.6 MB     Downloading: 5.7 MB     Downloading: 5.7 MB     Downloading: 5.7 MB     Downloading: 5.7 MB     Downloading: 5.8 MB     Downloading: 5.8 MB     Downloading: 5.8 MB     Downloading: 5.8 MB     Downloading: 5.8 MB     Downloading: 5.8 MB     Downloading: 5.9 MB     Downloading: 5.9 MB     Downloading: 6 MB     Downloading: 6 MB     Downloading: 6 MB     Downloading: 6 MB     Downloading: 6 MB     Downloading: 6 MB     Downloading: 6 MB     Downloading: 6 MB     Downloading: 6.1 MB     Downloading: 6.1 MB     Downloading: 6.1 MB     Downloading: 6.1 MB     Downloading: 6.1 MB     Downloading: 6.1 MB     Downloading: 6.2 MB     Downloading: 6.2 MB     Downloading: 6.2 MB     Downloading: 6.2 MB     Downloading: 6.2 MB     Downloading: 6.2 MB     Downloading: 6.3 MB     Downloading: 6.3 MB     Downloading: 6.3 MB     Downloading: 6.3 MB     Downloading: 6.4 MB     Downloading: 6.4 MB     Downloading: 6.4 MB     Downloading: 6.4 MB     Downloading: 6.5 MB     Downloading: 6.5 MB     Downloading: 6.5 MB     Downloading: 6.5 MB     Downloading: 6.5 MB     Downloading: 6.5 MB     Downloading: 6.5 MB     Downloading: 6.5 MB     Downloading: 6.6 MB     Downloading: 6.6 MB     Downloading: 6.6 MB     Downloading: 6.6 MB     Downloading: 6.6 MB     Downloading: 6.6 MB     Downloading: 6.7 MB     Downloading: 6.7 MB     Downloading: 6.7 MB     Downloading: 6.7 MB     Downloading: 6.8 MB     Downloading: 6.8 MB     Downloading: 6.8 MB     Downloading: 6.8 MB     Downloading: 6.9 MB     Downloading: 6.9 MB     Downloading: 6.9 MB     Downloading: 6.9 MB     Downloading: 6.9 MB     Downloading: 6.9 MB     Downloading: 7 MB     Downloading: 7 MB     Downloading: 7 MB     Downloading: 7 MB     Downloading: 7 MB     Downloading: 7 MB     Downloading: 7 MB     Downloading: 7 MB     Downloading: 7 MB     Downloading: 7 MB     Downloading: 7 MB     Downloading: 7 MB     Downloading: 7.1 MB     Downloading: 7.1 MB     Downloading: 7.1 MB     Downloading: 7.1 MB     Downloading: 7.2 MB     Downloading: 7.2 MB     Downloading: 7.2 MB     Downloading: 7.2 MB     Downloading: 7.3 MB     Downloading: 7.3 MB     Downloading: 7.3 MB     Downloading: 7.3 MB     Downloading: 7.3 MB     Downloading: 7.3 MB     Downloading: 7.3 MB     Downloading: 7.3 MB     Downloading: 7.4 MB     Downloading: 7.4 MB     Downloading: 7.4 MB     Downloading: 7.4 MB     Downloading: 7.4 MB     Downloading: 7.4 MB     Downloading: 7.4 MB     Downloading: 7.4 MB     Downloading: 7.6 MB     Downloading: 7.6 MB     Downloading: 7.6 MB     Downloading: 7.6 MB     Downloading: 7.6 MB     Downloading: 7.6 MB     Downloading: 7.6 MB     Downloading: 7.6 MB     Downloading: 7.9 MB     Downloading: 7.9 MB     Downloading: 7.9 MB     Downloading: 7.9 MB     Downloading: 7.9 MB     Downloading: 7.9 MB     Downloading: 7.9 MB     Downloading: 7.9 MB     Downloading: 8 MB     Downloading: 8 MB     Downloading: 8 MB     Downloading: 8 MB     Downloading: 8 MB     Downloading: 8 MB     Downloading: 8.1 MB     Downloading: 8.1 MB     Downloading: 8.1 MB     Downloading: 8.1 MB     Downloading: 8.1 MB     Downloading: 8.1 MB     Downloading: 8.1 MB     Downloading: 8.1 MB     Downloading: 8.1 MB     Downloading: 8.1 MB     Downloading: 8.1 MB     Downloading: 8.1 MB     Downloading: 8.2 MB     Downloading: 8.2 MB     Downloading: 8.2 MB     Downloading: 8.2 MB     Downloading: 8.2 MB     Downloading: 8.2 MB     Downloading: 8.2 MB     Downloading: 8.2 MB     Downloading: 8.3 MB     Downloading: 8.3 MB     Downloading: 8.3 MB     Downloading: 8.3 MB     Downloading: 8.3 MB     Downloading: 8.3 MB     Downloading: 8.3 MB     Downloading: 8.3 MB     Downloading: 8.4 MB     Downloading: 8.4 MB     Downloading: 8.4 MB     Downloading: 8.4 MB     Downloading: 8.4 MB     Downloading: 8.4 MB     Downloading: 8.4 MB     Downloading: 8.4 MB     Downloading: 8.4 MB     Downloading: 8.4 MB     Downloading: 8.5 MB     Downloading: 8.5 MB     Downloading: 8.5 MB     Downloading: 8.5 MB     Downloading: 8.5 MB     Downloading: 8.5 MB     Downloading: 8.5 MB     Downloading: 8.5 MB     Downloading: 8.5 MB     Downloading: 8.5 MB     Downloading: 8.6 MB     Downloading: 8.6 MB     Downloading: 8.6 MB     Downloading: 8.6 MB     Downloading: 8.6 MB     Downloading: 8.6 MB     Downloading: 8.7 MB     Downloading: 8.7 MB     Downloading: 8.7 MB     Downloading: 8.7 MB     Downloading: 8.8 MB     Downloading: 8.8 MB     Downloading: 8.8 MB     Downloading: 8.8 MB     Downloading: 8.8 MB     Downloading: 8.8 MB     Downloading: 8.8 MB     Downloading: 8.8 MB     Downloading: 8.8 MB     Downloading: 8.8 MB     Downloading: 8.9 MB     Downloading: 8.9 MB     Downloading: 8.9 MB     Downloading: 8.9 MB     Downloading: 8.9 MB     Downloading: 8.9 MB     Downloading: 8.9 MB     Downloading: 8.9 MB     Downloading: 8.9 MB     Downloading: 8.9 MB     Downloading: 9 MB     Downloading: 9 MB     Downloading: 9 MB     Downloading: 9 MB     Downloading: 9 MB     Downloading: 9 MB     Downloading: 9.1 MB     Downloading: 9.1 MB     Downloading: 9.1 MB     Downloading: 9.1 MB     Downloading: 9.1 MB     Downloading: 9.1 MB     Downloading: 9.1 MB     Downloading: 9.1 MB     Downloading: 9.1 MB     Downloading: 9.1 MB     Downloading: 9.1 MB     Downloading: 9.1 MB     Downloading: 9.1 MB     Downloading: 9.1 MB     Downloading: 9.1 MB     Downloading: 9.1 MB     Downloading: 9.1 MB     Downloading: 9.1 MB     Downloading: 9.2 MB     Downloading: 9.2 MB     Downloading: 9.2 MB     Downloading: 9.2 MB     Downloading: 9.2 MB     Downloading: 9.2 MB     Downloading: 9.3 MB     Downloading: 9.3 MB     Downloading: 9.3 MB     Downloading: 9.3 MB     Downloading: 9.3 MB     Downloading: 9.3 MB     Downloading: 9.3 MB     Downloading: 9.3 MB     Downloading: 9.3 MB     Downloading: 9.3 MB     Downloading: 9.3 MB     Downloading: 9.3 MB     Downloading: 9.4 MB     Downloading: 9.4 MB     Downloading: 9.4 MB     Downloading: 9.4 MB     Downloading: 9.4 MB     Downloading: 9.4 MB     Downloading: 9.5 MB     Downloading: 9.5 MB     Downloading: 9.5 MB     Downloading: 9.5 MB     Downloading: 9.5 MB     Downloading: 9.5 MB     Downloading: 9.5 MB     Downloading: 9.5 MB     Downloading: 9.5 MB     Downloading: 9.5 MB     Downloading: 9.5 MB     Downloading: 9.5 MB     Downloading: 9.6 MB     Downloading: 9.6 MB     Downloading: 9.6 MB     Downloading: 9.6 MB     Downloading: 9.6 MB     Downloading: 9.6 MB     Downloading: 9.6 MB     Downloading: 9.6 MB     Downloading: 9.6 MB     Downloading: 9.6 MB     Downloading: 9.6 MB     Downloading: 9.6 MB     Downloading: 9.7 MB     Downloading: 9.7 MB     Downloading: 9.7 MB     Downloading: 9.7 MB     Downloading: 9.7 MB     Downloading: 9.7 MB     Downloading: 9.7 MB     Downloading: 9.7 MB     Downloading: 9.7 MB     Downloading: 9.7 MB     Downloading: 9.8 MB     Downloading: 9.8 MB     Downloading: 9.8 MB     Downloading: 9.8 MB     Downloading: 9.8 MB     Downloading: 9.8 MB     Downloading: 9.8 MB     Downloading: 9.8 MB     Downloading: 9.8 MB     Downloading: 9.8 MB     Downloading: 9.8 MB     Downloading: 9.8 MB     Downloading: 9.9 MB     Downloading: 9.9 MB     Downloading: 9.9 MB     Downloading: 9.9 MB     Downloading: 9.9 MB     Downloading: 9.9 MB     Downloading: 9.9 MB     Downloading: 9.9 MB     Downloading: 9.9 MB     Downloading: 9.9 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 10 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 11 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 12 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 13 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     Downloading: 14 MB     

pl_retally(RI_2010, crosswalk)

#> # A tibble: 25,649 x 24
#>    GEOID     state county    vtd   area_land area_water   pop pop_hisp pop_white
#>    <chr>     <chr> <chr>     <chr>     <dbl>      <dbl> <int>    <int>     <int>
#>  1 44001030… RI    Bristol … NA         5329          0    15        3        13
#>  2 44001030… RI    Bristol … NA         6001          0     8        0         6
#>  3 44001030… RI    Bristol … NA        11821          0    17        0        17
#>  4 44001030… RI    Bristol … NA        16049          0    28        0        28
#>  5 44001030… RI    Bristol … NA         5398          0    12        0         9
#>  6 44001030… RI    Bristol … NA       181092          0    93        1        84
#>  7 44001030… RI    Bristol … NA        17056          0    23        0        23
#>  8 44001030… RI    Bristol … NA        16672          0    28        0        28
#>  9 44001030… RI    Bristol … NA         7291          0     9        0         9
#> 10 44001030… RI    Bristol … NA       164924          0    70        1        64
#> # … with 25,639 more rows, and 15 more variables: pop_black <int>,
#> #   pop_aian <int>, pop_asian <int>, pop_nhpi <int>, pop_other <int>,
#> #   pop_two <int>, vap <int>, vap_hisp <int>, vap_white <int>, vap_black <int>,
#> #   vap_aian <int>, vap_asian <int>, vap_nhpi <int>, vap_other <int>,
#> #   vap_two <int>

# }