Approximately re-tally Census data under new block boundaries

Applies a block crosswalk to a table of block data using areal interpolation. That is, the fraction of land area in the overlapping region between old and new blocks is used to divide the population of the old blocks into the new.

pl_retally(d_from, crosswalk)

Arguments

d_from
The data frame to process. All numeric columns will be re-tallied. Integer columns will be re-tallied with rounding. Character columns will be preserved if constant across new block geometries.

crosswalk
The crosswalk data frame, from pl_crosswalk()

Value

A new data frame, like d_from, except with the geometry column dropped, if one exists. New geometry should be loaded, perhaps with tigris::blocks().

Details

All numeric columns will be re-tallied. Integer columns will be re-tallied with rounding. Character columns will be preserved if constant across new block geometries.

Blocks from other states will be ignored.

Examples

# \donttest{
crosswalk = pl_crosswalk("RI", 2010, 2020)
RI_2010 = pl_tidy_shp("RI", pl_url("RI", 2010), 2010)
pl_retally(RI_2010, crosswalk)

#> # A tibble: 25,649 x 24
#>    GEOID     state county    vtd   area_land area_water   pop pop_hisp pop_white
#>    <chr>     <chr> <chr>     <chr>     <dbl>      <dbl> <int>    <int>     <int>
#>  1 44001030… RI    Bristol … NA         5329          0    15        3        13
#>  2 44001030… RI    Bristol … NA         6001          0     8        0         6
#>  3 44001030… RI    Bristol … NA        11821          0    17        0        17
#>  4 44001030… RI    Bristol … NA        16049          0    28        0        28
#>  5 44001030… RI    Bristol … NA         5398          0    12        0         9
#>  6 44001030… RI    Bristol … NA       181092          0    93        1        84
#>  7 44001030… RI    Bristol … NA        17056          0    23        0        23
#>  8 44001030… RI    Bristol … NA        16672          0    28        0        28
#>  9 44001030… RI    Bristol … NA         7291          0     9        0         9
#> 10 44001030… RI    Bristol … NA       164924          0    70        1        64
#> # … with 25,639 more rows, and 15 more variables: pop_black <int>,
#> #   pop_aian <int>, pop_asian <int>, pop_nhpi <int>, pop_other <int>,
#> #   pop_two <int>, vap <int>, vap_hisp <int>, vap_white <int>, vap_black <int>,
#> #   vap_aian <int>, vap_asian <int>, vap_nhpi <int>, vap_other <int>,
#> #   vap_two <int>

# }